requirements.txt
# barbican
# cinder
# glance
# keystone
# neutron
# nova
PyYAML
oslo.config
